refactor: tree.py: renamed from write_tree
